Don't show pages on inactive docs in sidebar
By default all the pages of all docs are listed in the sidebar as a h3
heading. This makes the sidebar nearly unreadable for large documents
and distracts too much. Up to a point that orienting in the document
structure gets quiet hard.

This commit changes this to only display the pages of the current doc
in the sidebar. All other docs will only be listed with the document
name (the _index.md from the sections folder).