Provide shortcode for feature boxes
Meant to be used on a landing or home page the feature boxes display an
icon with a caption and a description in a grid view.

While the description can be long and contain markdown it should be kept
terse, more like a catch-phrase.

The icon along with the caption will be a link to some (internal or
external) web page.